Update error handling example in test/compiler.js
The current example results in an unhelpful stack trace when `stats.hasErrors() === true`:
```
TypeError: Cannot read property 'message' of null

      at isError (node_modules/jest-jasmine2/build/is_error.js:20:27)
```
This is because we are always rejecting with `err`. Plucked and modified this example from [Error Handling](https://webpack.js.org/api/node/#error-handling) docs and it results in a more useful stack trace.